refactor: unify FreeBSD and OpenBSD backends into a single bsd.go file
